kana-writing/requirements.txt
# Backend
fastapi
uvicorn
python-multipart>=0.0.6
pytest
httpx>=0.25.1
manga-ocr>=0.1.8

# Frontend
streamlit
pillow
streamlit-drawable-canvas
requests
